- moved compiler detection in SETTINGS section using DETECT option. - fixed bad stuff discovered by lint (wasn't changing pmk's behavior). - added some data in pmkcomp.dat about shared libs.

+++ hash.h
@@ -81,7 +81,7 @@ typedef struct {
 } hpair;
 
 typedef struct {
-	int	  nkey;
+	size_t	  nkey;
 	char	**keys;
 } hkeys;
